ERR_EMPTY_RESPONSE on the /channels
Analysis of the Error
The Cause: The error is in the channels_page procedure within the arm-24_web.tcl script.

The Trigger: When building the page, the code attempts to retrieve the mode, url, and desc settings for every channel. However, if a channel has never had one of these settings configured, it won't have an entry in the database. The code tries to get a value that doesn't exist, which causes a Tcl error and crashes the procedure.

The Solution
The fix is to make the channels_page procedure more resilient by checking if a setting exists before trying to display it.